added app module, injected manager classes
