+++ app/src/main/java/com/yildirimomer01/ewa/util/AssetManager.kt
@@ -1,0 +1,7 @@
+package com.yildirimomer01.ewa.util
+
+import android.content.Context
+
+interface AssetManager {
+    fun getJsonAsset(context: Context, fileName: String): String
+}
