+++ app/src/main/java/com/yildirimomer01/ewa/util/AssetManagerImpl.kt
@@ -5,7 +5,7 @@ import android.content.Context
 /**
  * class that manages asset operations
  */
-class MockAssetManager() {
-    fun getJsonAsset(context: Context, fileName: String) =
+class AssetManagerImpl : AssetManager {
+    override fun getJsonAsset(context: Context, fileName: String) =
         context.assets.open(fileName).bufferedReader().use { it.readText() }
 }
